Add ability to register a user
NOTE: This is hardcoded right now to the username/password style, which
is not agnostic enough to support all xmpp use cases.  However, it
supports the default for ejabberd and mongooseim, which is at least a
decent start.  Unsure if it should be merged or if we should support
something more generic here instead.