Make console test more reliable + pass in timeout
By using a `Task` for running the `send_to_adapter` call, we can avoid an
arbitrary sleep timer where we *hope* the process is done. In addition,
add the timeout used in `await` as an optional parameter, so that we
can pass in a shorter timeout when running the test.